# docs/openapi-v1.yml
openapi: 3.0.3
info:
  title: Selene API — v1
  description: >
    Public REST contract for the Selene smart energy & climate platform.
    Served by the Caddy gateway under the `/api/v1` prefix; each path below is
    routed to a dedicated microservice (services/auth, services/energy,
    services/climate, services/analytics, services/firmware).

    Versioning lives at the gateway: `handle_path /api/v1/*` strips the prefix
    before proxying, so services define version-agnostic paths (`/energy/latest`)
    and a future `/api/v2` needs no service changes.

    Auth: short-lived EdDSA (Ed25519) access tokens with rotating refresh
    tokens (one-time use, reuse detection). JWKS at `/.well-known/jwks.json`.
  version: 1.0.0
  contact:
    name: Selene
  license:
    name: Proprietary
    url: https://github.com/dankehidayat/selene
servers:
  - url: https://selene.dankehidayat.my.id/api/v1
    description: Production
  - url: http://localhost:3009/api/v1
    description: Local (gateway / auth)
...
paths:
  /auth/register:
    post:

      operationId: registerUser
      tags: [auth]
      summary: Create an account
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Account created; access + refresh tokens issued.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400": { $ref: "#/components/responses/BadRequest" }
        "409": { $ref: "#/components/responses/Conflict" }
  /auth/login:
    post:

      operationId: login
      tags: [auth]
      summary: Sign in (may require a 2FA step)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Signed in, or 2FA required in `requires2fa` mode.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /auth/login/2fa:
    post:

      operationId: loginTwoFactor
      tags: [auth]
      summary: Complete login with TOTP or backup code
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TwoFactorLoginRequest"
      responses:
        "200":
          description: Authenticated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /auth/refresh:
    post:

      operationId: refreshTokens
      tags: [auth]
      summary: Rotate an access token using a refresh token (one-time use)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: New token pair; the previous refresh token is invalidated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RefreshResponse"
        "401":
          description: Invalid or reused refresh token (reuses force logout).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /auth/logout:
    post:

      operationId: logout
      tags: [auth]
      summary: Revoke the current session
      security: [{ bearerAuth: [] }]
      responses:
        "204": { description: Revoked }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /me:
    get:

      operationId: getProfile
      tags: [me]
      summary: Current user profile
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Profile.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401": { $ref: "#/components/responses/Unauthorized" }
    patch:
      tags: [me]
      summary: Update profile (name, etc.)
      operationId: updateProfile
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: Profile updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401": { $ref: "#/components/responses/Unauthorized" }
  /me/password:
    post:

      operationId: changePassword
      tags: [me]
      summary: Change password (invalidates other sessions)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequest"
      responses:
        "204":
          description: Password changed.
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /auth/forgot-password:
    post:

      operationId: forgotPassword
      tags: [auth]
      summary: Request a password reset (always generic response)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordRequest"
      responses:
        "200":
          description: Always returns the generic success message.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  message:
                    type: string
        "429": { $ref: "#/components/responses/RateLimited" }
  /auth/reset-password:
    post:

      operationId: resetPassword
      tags: [auth]
      summary: Complete a password reset with the emailed token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordRequest"
      responses:
        "204":
          description: Password reset.
        "400":
          description: Invalid or expired token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short-lived EdDSA (Ed25519) access token; refresh via POST /auth/refresh. Verify against /.well-known/jwks.json.
...
  responses:
    BadRequest:
      description: Invalid payload.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Unauthorized:
      description: Missing, invalid, or expired token.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    Conflict:
      description: Already exists (e.g. email).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    RateLimited:
      description: Too many requests; honor the Retry-After header.
      headers:
        Retry-After:
          schema: { type: integer }
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message]
          properties:
            code:
              type: string
              enum: [UNAUTHORIZED, FORBIDDEN, RATE_LIMITED, VALIDATION_ERROR, NOT_FOUND, CONFLICT, INTERNAL_ERROR, TOKEN_REUSE_DETECTED]
            message:
              type: string
            details:
              type: object
              additionalProperties: true
            timestamp:
              type: string
              format: date-time
            request_id:
              type: string
    RegisterRequest:
      type: object
      additionalProperties: false
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
          minLength: 6
        name:
          type: string
    LoginRequest:
      type: object
      additionalProperties: false
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    TwoFactorLoginRequest:
      type: object
      additionalProperties: false
      required: [pendingToken, code]
      properties:
        pendingToken:
          type: string
        code:
          type: string
          description: TOTP or backup code.
    RefreshRequest:
      type: object
      additionalProperties: false
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
    RefreshResponse:
      type: object
      additionalProperties: false
      required: [accessToken, refreshToken, expiresIn]
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        expiresIn:
          type: integer
          description: Access-token lifetime in seconds (e.g. 900).
    LoginResponse:
      oneOf:
        - $ref: "#/components/schemas/AuthResponse"
        - $ref: "#/components/schemas/TwoFactorChallenge"
    AuthResponse:
      type: object
      additionalProperties: false
      required: [accessToken, refreshToken, user]
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        expiresIn:
          type: integer
        user:
          $ref: "#/components/schemas/UserPublic"
    TwoFactorChallenge:
      type: object
      additionalProperties: false
      required: [requires2fa, pendingToken]
      properties:
        requires2fa:
          type: boolean
        pendingToken:
          type: string
        message:
          type: string
...
    UserPublic:
      type: object
      additionalProperties: false
      required: [id, email, role]
      properties:
        id: { type: string }
        email: { type: string, format: email }
        name: { type: string, nullable: true }
        role:
          type: string
          enum: [USER, ADMIN]
        totpEnabled:
          type: boolean
        createdAt:
          type: string
          format: date-time
    UserProfile:
      type: object
      additionalProperties: false
      required: [id, email, role]
      properties:
        id: { type: string }
        email: { type: string }
        name: { type: string, nullable: true }
        role: { type: string, enum: [USER, ADMIN] }
        totpEnabled: { type: boolean }
        createdAt: { type: string, format: date-time }
    UpdateProfileRequest:
      type: object
      additionalProperties: false
      properties:
        name: { type: string }
    ChangePasswordRequest:
      type: object
      additionalProperties: false
      required: [currentPassword, newPassword]
      properties:
        currentPassword: { type: string, format: password }
        newPassword:
          type: string
          format: password
          minLength: 6
    ForgotPasswordRequest:
      type: object
      additionalProperties: false
      required: [email]
      properties:
        email: { type: string, format: email }
    ResetPasswordRequest:
      type: object
      additionalProperties: false
      required: [token, newPassword]
      properties:
        token: { type: string }
        newPassword: { type: string, format: password, minLength: 6 }
...